refactor(dashboard): split view into colocated section directories
Break dashboard/index.vue into a thin orchestrator with member-section,
deck-grid, and review-inbox each owning their own directory and
sub-components, matching the rest of the codebase's view layout.
