+++ src/views/dashboard/review-inbox/index.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-import ReviewInboxItem from './review-inbox-item.vue'
-import ReviewInboxNavButton from './review-inbox-nav-button.vue'
+import ReviewInboxItem from './item.vue'
+import ReviewInboxNavButton from './nav-button.vue'
 import { useStudyModal } from '@/views/study-session/composables/study-modal'
-import { useReviewInboxScroll } from './use-review-inbox-scroll'
+import { useReviewInboxScroll } from './use-scroll'
 
 type ReviewInboxProps = {
   due_decks: Deck[]
@@ -20,10 +20,7 @@ function onItemClicked(deck: Deck) {
 </script>
 
 <template>
-  <div
-    data-testid="review-inbox"
-    class="w-full rounded-8 bg-brown-300 dark:bg-stone-900 select-none flex overflow-hidden"
-  >
+  <div data-testid="review-inbox" class="w-full select-none flex overflow-hidden">
     <review-inbox-nav-button
       v-if="has_overflow"
       direction="prev"
@@ -34,8 +31,7 @@ function onItemClicked(deck: Deck) {
     <div
       ref="items_el"
       data-testid="review-inbox__items"
-      class="flex-1 min-w-0 flex gap-1 overflow-x-auto overflow-y-hidden snap-x snap-mandatory scroll-px-4 touch-pan-x scroll-hidden p-4"
-      :class="[has_overflow ? 'rounded-8' : 'justify-center']"
+      class="flex-1 min-w-0 flex gap-3 overflow-x-auto overflow-y-hidden snap-x snap-mandatory scroll-px-4 touch-pan-x scroll-hidden pt-4"
     >
       <review-inbox-item
         v-for="deck in due_decks"
